refactor: 총구매액 계산 로직 이동
Payment가 아닌 Order에게 책임을 넘긴다.
주문이 재고에 의존하여 계산한다.
